requirements.txt
blinker==1.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.19.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
WTForms==2.3.3
